Fix a harmless compiler warning in tclsqlite.c for MSVC x64.  Fix a test case associated with the FTS4 merge feature.
FossilOrigin-Name: de3f7187eb2ff24683a1854ae523e7d30c290893

+++ src/tclsqlite.c
@@ -1163,7 +1163,7 @@ static int dbPrepareAndBind(
     memset(pPreStmt, 0, nByte);
 
     pPreStmt->pStmt = pStmt;
-    pPreStmt->nSql = (*pzOut - zSql);
+    pPreStmt->nSql = (int)(*pzOut - zSql);
     pPreStmt->zSql = sqlite3_sql(pStmt);
     pPreStmt->apParm = (Tcl_Obj **)&pPreStmt[1];
 #ifdef SQLITE_TEST
